[test] Refactor adding DescribeExecute for testing executor with and without concurrent.Executor.
